requirements.txt
FastAPI==0.95.0
uvicorn==0.22.0
httpx==0.24.0
pydantic==1.10.7
python-multipart==0.0.6
aiofiles==0.8.0
openai==1.64.0
python-dotenv==1.0.1